Put enemies into map

+++ Assets/Scripts/Gameplay/Weapon.cs
@@ -8,7 +8,7 @@ public class Weapon : MonoBehaviour
     [SerializeField] protected float speed;
     [SerializeField] protected int numProjectiles;
     protected int curAmmoCount;
-    [SerializeField] protected int reloadTime;
+    [SerializeField] protected float reloadTime;
     protected bool canShoot;
 
     protected virtual void fire()
